ajout de sécurité

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 
 exports.createThing = (req, res, next) => {
@@ -18,16 +19,42 @@ exports.createThing = (req, res, next) => {
 
 
 exports.modifyThings = (req, res, next) => { 
-    Thing.updateOne({ _id: req.params.id}, {...req.body, _id: req.params.id})
-        .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
-        .catch(error => res.status(400).json({ error }));
-}
+    const thingObject = req.file ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+    delete thingObject._userId;
+    Thing.findOne({_id: req.params.id})
+        .then((thing)=> {
+            if(thing.userId != req.auth.userId){
+                res.status(401).json({ message: 'Non-autorisé'});
+            }else{
+                Thing.updateOne({ _id: req.params.id}, {...thingObject, _id: req.params.id})
+                    .then(() => res.status(200).json({ message: 'Objet modififé !'}))
+                    .catch((error)=> res.status(401).json({ error }));
+            }
+        })
+        .catch((error)=> res.status(400).json({ error }));
 
-exports.deleteThings = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id})
-        .then(() => res.status(200).json({ message: 'Sauce suprimée !'}))
-        .catch(error => res.status(400).json({ error }));
-}
+};
+exports.deleteThing = (req, res, next) => {
+    Thing.findOne({ _id: req.params.id})
+        .then(thing => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error })
+        });
+ };
 
 exports.getOneThing = (req, res, next) => { 
     Thing.findOne({ _id: req.params.id })
